Restructure inventory package
