+++ src/main/java/com/thebois/models/inventory/items/ItemType.java
@@ -1,4 +1,4 @@
-package com.thebois.models.inventory;
+package com.thebois.models.inventory.items;
 
 /**
  * All the different types of items in the game.
